Tf-simple-plan function feedback when filtering
Switch from tee to iterate over each line of sdtin, allowing to give the current line as feedback of the command running, but always cleaning the previous line to avoid flood on the terminal. That improves the feedback of the command where before it was confusing about if some error happened or if it is even running at all. The feedback will still avoid printing everything on the screen by overriding the same line multiple times and removing it at the end.