tf-simple-plan finish feedback when pipes end
Avoid printing each line of the tf diffs since it will mean the upstream pipe ends. The terminal print is a slow operation, and holding on to it when it is already done wastes time and resources. Also, it makes a significant impact when the diff is big where the user waits many seconds for something already done, but now it will detect and end faster.