Remove unused apps from Brewfile
Some are still usuful but not in my currenct context. I commented the ones that I plan to use in future with a "maybe" in some cases.
I removed the ones that I'll probably not use, or that are very scoped to work needs that I don't have plans in short/mid-term.

I'm also trying to reduce the number of things installed. My old computer is getting degradation performance, not just for this, but for apps that are nice, but heavy.